Use search field for search.
Search options are presented in a context menu on the search field, as well as
links in search header. A 'header' is still displayed while in search mode, to
clearly indicate the mode.

This provides a slightly simpler (no extra click), more common UI, while keeping
the all the features of the earlier search mode.