Add hint for Klipper to not add passwords to history
Keepassx offers an option to clear the clipboard/selection
after some time, e.g. 10 seconds, after the password was
copied to the clipboard. This works fine, but unfortunately
the password isn't removed from Klipper's history. This is
a great security risk, which may make the use of keepassx
impossible.

This patch adds the additional mime type
'x-kde-passwordManagerHint' to the mimeData when copying a
password to the clipboard. Klipper will then not insert it
into its history.

See https://phabricator.kde.org/D12539