unhide the keepassx window after autotype is done
When pressing Ctrl+V to autotype the password, keepassx window is
hidden(minimized). This patch unhides it after autotype is finished.
This allows for just Alt+Tab to switch back to keepassx after autotype
is done, rather than having to find keepassx in taskbar and click it to
restore its window. (time consuming)
This unhide doesn't automatically focus keepassx though, or it's my Xfce
window manager settings that prevent focus stealing if it otherwise would.